+++ src/main/java/com/Utp/DesarrolloWeb/controller/ProductoController.java
@@ -3,8 +3,8 @@ package com.Utp.DesarrolloWeb.controller;
 import com.Utp.DesarrolloWeb.model.Producto;
 import com.Utp.DesarrolloWeb.service.ProductoService;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
+
 @RestController
 @RequestMapping("/productos")
 public class ProductoController {
@@ -26,17 +26,17 @@ public class ProductoController {
     }
 
     @GetMapping("/{id}")
-    public Producto buscar(@PathVariable int id) {
+    public Producto buscar(@PathVariable Long id) {
         return servicio.buscarPorId(id);
     }
 
     @DeleteMapping("/{id}")
-    public String eliminar(@PathVariable int id) {
+    public String eliminar(@PathVariable Long id) {
         return servicio.eliminar(id);
     }
 
     @PutMapping("/{id}")
-    public Producto actualizar(@PathVariable int id, @RequestBody Producto p) {
+    public Producto actualizar(@PathVariable Long id, @RequestBody Producto p) {
         return servicio.actualizar(id, p);
     }
 }
